+++ src/shared/schemas.py
@@ -23,6 +23,10 @@ class HTTPError(Schema):
     details: str
 
 
+class NotAuthenticatedResponse(HTTPError):  # noqa: N818
+    """Schema for 401 UNAUTHORIZED response."""
+
+
 class NotFoundResponse(HTTPError):  # noqa: N818
     """Schema for 404 NOT FOUND error."""
