+++ src/shared/swagger.py
@@ -8,6 +8,18 @@ from src.shared import exceptions, schemas
 
 
 responses = {  # pylint: disable=consider-using-namedtuple-or-dataclass
+    status.HTTP_401_UNAUTHORIZED: {
+        "description": "Provided tokens or credentials are invalid or missing.",
+        "model": schemas.NotAuthenticatedResponse,
+        "content": {
+            "application/json": {
+                "example": {
+                    "description": exceptions.NotAuthenticatedException.description,
+                    "details": exceptions.NotAuthenticatedException.details,
+                },
+            },
+        },
+    },
     status.HTTP_403_FORBIDDEN: {
         "description": "Provided tokens or credentials don't grant you enough access rights.",
         "model": schemas.NotAllowedResponse,
